commit anterior pt 2

+++ refactor/src/utils/value_utils.h
@@ -234,10 +234,10 @@ namespace ValueUtils {
     template<typename T, typename compareLT>
     void heapSort(std::vector<T>& arr, compareLT lt, bool ascending = true){
         size_t num_elements = arr.size();
-        for (size_t i = num_elements / 2 - 1; i >= 0; i--) {
+        for (int i = num_elements / 2 - 1; i >= 0; i--) {
             heapify(arr, num_elements, i, lt, ascending);
         }
-        for (size_t i = num_elements - 1; i > 0; i--) {
+        for (int i = num_elements - 1; i > 0; i--) {
             std::swap(arr[0], arr[i]);
             heapify(arr, i, 0, lt, ascending);
         }
@@ -246,10 +246,10 @@ namespace ValueUtils {
     template<typename Iterator, typename compareLT>
     void heapSort(Iterator begin, Iterator end, compareLT lt, bool ascending = true) {
         size_t num_elements = std::distance(begin, end);
-        for (size_t i = num_elements / 2 - 1; i >= 0; i--) {
+        for (int i = num_elements / 2 - 1; i >= 0; i--) {
             heapify(begin, end, num_elements, i, lt, ascending);
         }
-        for (size_t i = num_elements - 1; i > 0; i--) {
+        for (int i = num_elements - 1; i > 0; i--) {
             std::swap(*begin, *(begin + i));
             heapify(begin, end, i, 0, lt, ascending);
         }
